Bug 833624. Bump up size of int's in mOrder to allow more than 256 CSS properties. r=dbaron
UltraBlame original commit: e7b55316545e330ee7ed2664b928f03f0a49ac6a

+++ layout/style/Declaration.h
@@ -230,7 +230,7 @@ public:
   size_t SizeOfIncludingThis(nsMallocSizeOfFun aMallocSizeOf) const;
 
 private:
-  nsAutoTArray<uint8_t, 8> mOrder;
+  nsAutoTArray<uint16_t, 8> mOrder;
